perf: split 'filter-products'
